Add 'list' parameter to list supported package types
We add an explicit 'list' value to the type check so that users can determine
the supported packages in an error-free fashion, per issue #10. Making the
parameter optional is technically possible, but with the number of issues
that GLib's GOption* API present, it seems likely it will be redone in the
future to be less insane.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>