template: Move to block-based render passes for sections
We now build a complete unexpanded section at a time whilst iterating the
document structure. As we exit each subsection, we push our block into the
parent render block. This leads to a bottom up render approach for emitting
sections, skipping non-emitters. In future this can be expanded upon by
enabling iterative conditional sections, as a full section block is available
to render as many times as required into the parent context.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>