Extend advisory locking support to serialize all access to database
One of the obvious reason for this change is to ensure check is
always performed against the same set of the CVE data in SQLite3
database, excluding the case where one instance updates database
and another uses it for CVE checks.

This behaviour is observed with Image Security Analyser Framework
(ISAFW) running with bitbake on YOCTO project builds in case where
no database exists or it's content is outdated, first instance
starts database update while other instances use database at the
same time, which they consided as up-to-date (i.e. with
update_required() returning false), for package checking. This
leads for incorrect results.

The approach to address descibled issues and other possible is to
extend advisory file locking usage to shared locks for read
database access and exclusive locking for write access, so
database updates only possible when there is no readers and only
one writer.

Introduce and use common API to work with advisory locking in both
cve-check-tool and cve-check-update. Now if there are readers and
database is out-of-date next instance can not aquire write lock
for update it waits to aquire it. If there are single writer all
reader instances can not aquire read lock they will wait for
writer to complete the update.

Lock timeout is kept from original database update locking
implementation, while adding command line argument option to
overwrite it at runtime might be useful.

Signed-off-by: Sergey Popovich <popovich_sergei@mail.ua>