update: Introduce NVD data integrity checks
National Vulnerability Database provides vulnerability information
in the form of XML files. These files are available in compressed
GZ or ZIP forms via HTTP/HTTPS protocol.

Fetching of the NVD XML could be incomplete for various reasons
(e.g. network reachability issues, interrupted download, etc).

To address this NVD provides META files for each XML data file.
This metadata provides last modification date/time, size of the
data in compressed/uncompressed forms and sha256 digest of the
uncompressed data. File is regular, plain text file with two
fields separated by the colon. First field must not contain
colon (as it is used as field separator), but second may contain it.

Introduce routines to parse NVD META files to get sha256 digest.
Compute digest on the fetched/unpacked data and compare it with
one from NVD META file.

If NVD XML files fetched we validate their integrity by calculating
sha256 on the uncompressed NVD XML file and comparing the result
with value from NVD META sha256 field.

If decompression fails or digest does not match we try to fetch NVD
data again and check the result, if that fails again we raise error.

If fetching of NVD file(s) was skipped (e.g. because no modification
is done to the file(s) on site since their local mtime value) and
data integrity isn't ok we try to uncompress and check digest again,
if that fails we try to refetch and check again.

While it might be reasonable to keep at least compressed XML
file and modify fetcher to resume failed downloads, this is
not preffered as solution due to the bugs in cURL related
to the HTTP 416 Requested Range Not Satisfiable error handling,
so make any failed downloads to start from the beginning.

Signed-off-by: Sergey Popovich <popovich_sergei@mail.ua>