util.c: Change logic in cve_get_file_parent()
Function cve_get_file_parent() will try to get the
realpath and the get the dirname. If the file used
to get parent doesn't exist the call will fail.

This problem is present when using another directory
for the database, realpath() won't find the nvd.db
file and the program will exit quitely.

This patch will first get the dirname and the get
the realpath to avoid failing when the doesn't exist.

Signed-off-by: Mariano Lopez <mariano.lopez@linux.intel.com>