Enable relative plugin locations
This allows finer integration into ISAFW, in that cve-check-tool will be
installed in a "syroot" mode, not a "host" mode. In this mode, cve-check-tool
will search for the directories ../lib/cve-check-tool and ../lib64/cve-check-tool
relative to it's main binary for the modules, in addition to the normal
host directory.

Note that this behaviour is disabled by default, as it's not meant to be used
in native-installed deployments due to the nature of forming .so loading paths
relative to /proc/self/exe.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>