Add initial revision of the template parser
This template parser implements a small subset of the Mustache template
specification, using {{ and }} to denote tags. These are used in most
cases as simple substitution variables.

Sections and nested sections are supported through child contexts. As such
adding a child context of name "bob" to the root context, would enable accessing
a section via the "{{#bob}}" tag. All sections must be accompanied by a closing
section tag, i.e. "{{/bob}}".

Sections can also be inverted, by specifying "^" before the tag name, i.e.
"{{^bob}}" in place of the standard "#". Also note that list types are supported
along with child contexts, so to add a list item requires adding to the list
with a child context, which becomes the new root node.

Parent traversal is also supported, in that global variables are available
in a child context, i.e. in a loop iteration. The reverse is not true, as
the subcontext variables are private, and not accessible by a parent or the
root context.

Currently no special handling is given to HTML strings to make them markup
safe, but this will land in a later revision, along with the ability to change
the mustache tags at runtime. Functions will also come along to make it
easier to deal with mmap()'d files, given that a design goal is to have a
persistently compiled template definition with the absolute minimum in
string duplication, to save on memory.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>