nvd: Vastly simplify things with bitwise tricks
Previously the design wasn't scalable to implementing other portions of the
NVD XML implementation, due to many boolean flags and conditional branches
in various places.

With this change, we can track a global table of tag -> mask mapping, and
then infer the logical flow of the document due to the flags currently set.
This also enables us to exactly match the document flow, whereas the previous
design required much manual logic to handle various parent container situations.

In short, this makes long term maintainence much easier, and was inspired in
part by sol.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>