update: Do not treat sha256 failure as fatal if requested
The NVD database is known to have issues during early morning whereby
the meta files don't actually match the sha256 of the target xml feed.
This has caused problems for some users of cve-check-tool, so in this
case we will now continue as if nothing fatal had happened if we find
the CVE_SKIP_VERIFY variable in the environment.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>